+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Entity Condition Types
+- Added `space` field to `velocity` entity condition type.
+
 ## Bugfixes
 - Fixed `compare_scales` crashing when `scale_type` is specified.
 - Fixed `compare_scales`s `scale_types` field not being counted.